fix: Move ops tests to board_tests.rs and fix draw_ops_mini_channel doc comment
Per york's review:
- Move is_ops_message and draw_ops_mini_channel tests from inline
  mod tests block to separate board_tests.rs per CLAUDE.md convention
- Fix stale doc comment on draw_ops_mini_channel to say "coworker
  /me action messages" instead of "/me action messages"
