+++ src/bin/midtown/cli/chat/ui/board.rs
@@ -466,7 +466,7 @@ fn render_task_item(
 
 /// Draw the Midtown Ops mini-channel (bottom of board sidebar).
 ///
-/// Displays system messages and /me action messages in a compact scrollable view.
+/// Displays system messages and coworker /me action messages in a compact scrollable view.
 /// The most recent `ops_messages` entries are shown (already in chronological order).
 fn draw_ops_mini_channel(f: &mut Frame, ops_messages: &[&midtown::Message], area: Rect) {
     // Split: header line + content
@@ -1685,129 +1685,6 @@ mod tests {
         );
     }
 
-    // --- is_ops_message tests ---
-
-    #[test]
-    fn test_is_ops_message_midtown_sender() {
-        assert!(is_ops_message("midtown", &MessageType::Text, "hello"));
-        assert!(is_ops_message("MIDTOWN", &MessageType::Text, "hello"));
-    }
-
-    #[test]
-    fn test_is_ops_message_github_sender() {
-        assert!(is_ops_message("github", &MessageType::System, "CI passed"));
-    }
-
-    #[test]
-    fn test_is_ops_message_system_sender() {
-        assert!(is_ops_message("system", &MessageType::System, "restart"));
-        assert!(is_ops_message("daemon", &MessageType::System, "spawned"));
-    }
-
-    #[test]
-    fn test_is_ops_message_action_type() {
-        // Action type (MessageType::Action) from coworkers is an ops message
-        assert!(is_ops_message(
-            "york",
-            &MessageType::Action,
-            "/me developing"
-        ));
-        // lead and user action messages stay in main chat
-        assert!(!is_ops_message(
-            "lead",
-            &MessageType::Action,
-            "/me reviewing"
-        ));
-    }
-
-    #[test]
-    fn test_is_ops_message_slash_me_content() {
-        // /me prefix in content = ops message even with Text type
-        assert!(is_ops_message("park", &MessageType::Text, "/me idle"));
-    }
-
-    #[test]
-    fn test_is_ops_message_regular_conversation() {
-        // Regular conversation from coworkers/lead/user is NOT ops
-        assert!(!is_ops_message("york", &MessageType::Text, "Hello team"));
-        assert!(!is_ops_message("lead", &MessageType::Text, "Looks good"));
-        assert!(!is_ops_message(
-            "user",
-            &MessageType::Text,
-            "What's the status?"
-        ));
-    }
-
-    #[test]
-    fn test_is_ops_message_lead_user_slash_me_not_ops() {
-        // lead and user /me actions stay in main chat, not ops
-        assert!(!is_ops_message(
-            "lead",
-            &MessageType::Action,
-            "/me reviewing"
-        ));
-        assert!(!is_ops_message(
-            "lead",
-            &MessageType::Text,
-            "/me checking things"
-        ));
-        assert!(!is_ops_message(
-            "user",
-            &MessageType::Action,
-            "/me asking a question"
-        ));
-        assert!(!is_ops_message(
-            "user",
-            &MessageType::Text,
-            "/me doing something"
-        ));
-    }
-
-    #[test]
-    fn test_draw_ops_mini_channel_renders_without_panic() {
-        use midtown::{Message, MessageType};
-        use ratatui::Terminal;
-        use ratatui::backend::TestBackend;
-
-        let backend = TestBackend::new(40, 8);
-        let mut terminal = Terminal::new(backend).unwrap();
-
-        let msgs: Vec<midtown::Message> = vec![
-            {
-                let mut m = Message::system("CI checks passed");
-                m.from = "github".to_string();
-                m
-            },
-            Message::new("york", "/me developing task 1583", MessageType::Action),
-        ];
-        let refs: Vec<&midtown::Message> = msgs.iter().collect();
-
-        terminal
-            .draw(|f| {
-                let area = f.area();
-                draw_ops_mini_channel(f, &refs, area);
-            })
-            .unwrap();
-        // No panic = success
-    }
-
-    #[test]
-    fn test_draw_ops_mini_channel_empty_messages() {
-        use ratatui::Terminal;
-        use ratatui::backend::TestBackend;
-
-        let backend = TestBackend::new(40, 5);
-        let mut terminal = Terminal::new(backend).unwrap();
-
-        terminal
-            .draw(|f| {
-                let area = f.area();
-                draw_ops_mini_channel(f, &[], area);
-            })
-            .unwrap();
-        // Empty state renders without panic
-    }
-
     #[test]
     fn test_draw_board_panel_includes_ops_section() {
         use ratatui::Terminal;
@@ -1844,3 +1721,7 @@ mod tests {
         );
     }
 }
+
+#[path = "board_tests.rs"]
+#[cfg(test)]
+mod board_tests;
